Reconsolidate environment variable handling
